Initial

+++ database/migrations/2023_01_24_220333_create_products_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('name');
             $table->string('slug');
             $table->integer('stock')->default(0);
-            $table->decimal('price',14);
+            $table->integer('price');
             $table->text('detail');
             $table->foreign('brand_id')->references('id')->on('brands');
             $table->timestamps();
