+++ database/migrations/2023_01_31_012656_create_cargos_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('cargos', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('price',5);
+            $table->integer('price');
             $table->tinyInteger('status')->default(1);
             $table->timestamps();
         });
